feat: replace getUserServersData with direct database fetching and status enrichment
